requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
psutil>=5.9.0
python-dateutil>=2.8.2 